+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [4.5.3] - 2022-05-18
+### Fixed
+- Simplified code of resource table component and components that use it.
+
 ## [4.5.2] - 2022-05-10
 ### Fixed
 - valueString Observations are not handled.